requirements.txt
flask>=2.0
tensorflow>=2.5
numpy>=1.19
pandas>=1.1
scikit-learn>=0.24
matplotlib>=3.3
opencv-python>=4.5
pillow>=8.0
jupyterlab>=3.0
